backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles==24.1.0",
    "fastapi==0.115.0",
    "fastapi-cli>=0.0.16",
    "gunicorn==23.0.0",
    "huggingface-hub==0.26.0",
    "pdfplumber==0.11.4",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "uvicorn[standard]==0.32.0",
]
